add isBrowser swtich to canvas related functions
